Refactor constant and move config into the folder.
