+++ src/constants/AuthorConstants.js
@@ -1,0 +1,5 @@
+'use strict';
+
+export default {
+  AUTHORS_EVENT: 'AUTHORS_EVENT'
+};
