+++ src/constants/MapConstants.js
@@ -1,0 +1,6 @@
+'use strict';
+
+export default {
+  MAP_EVENT: 'MAP_EVENT',
+  MAP_URL: '/data/africa.geo.json'
+};
