+++ src/constants/NationConstants.js
@@ -1,0 +1,8 @@
+'use strict';
+
+export default {
+  NATION_UPDATE: 10,
+  NATION_LOAD_DATA: 11,
+
+  NATIONS_SHOW_EVENT: 'NATIONS_SHOW_EVENT'
+};
