+++ src/constants/RouteConstants.js
@@ -2,11 +2,6 @@
 
 export default {
   ROUTE_UPDATE_PATH: 1,
-  NATION_UPDATE: 10,
-  NATION_LOAD_DATA: 11,
 
-  ROUTE_EVENT: 'ROUTE_EVENT',
-  AUTHORS_EVENT: 'AUTHORS_EVENT',
-  MAP_EVENT: 'MAP_EVENT',
-  NATIONS_SHOW_EVENT: 'NATIONS_SHOW_EVENT'
+  ROUTE_EVENT: 'ROUTE_EVENT'
 };
